refactor scheduler as a prop for centralized data flow
