+++ src/components/admin/tabs/SchedulesTab.tsx
@@ -10,35 +10,133 @@ import { Label } from "@/components/ui/label";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
+import { useState } from "react";
 import { Schedule } from "@/types";
+import * as adminService from "@/app/admin/adminService"
 
-//hooks
-import { useSchedules } from "@/hooks/useSchedules";
+interface ScheduleProp {
+    schedules: Schedule[];
+    fetchSchedules: () => Promise<void>;
+}
 
-export function SchedulesTab() {
-  
-  // extract tanan states ug handlers gikan sa hook 
-  const {
-    schedules,
-    newDateInput, setNewDateInput,
-    sessionType, setSessionType,
-    newAmCapacity, setNewAmCapacity,
-    newPmCapacity, setNewPmCapacity,
-    isAddDateOpen, setIsAddDateOpen,
-    manualStudentId, setManualStudentId,
-    isAddStudentOpen, setIsAddStudentOpen,
-    activeAddStudentSession, 
-    isEditCapacityOpen, setIsEditCapacityOpen,
-    editingCapacity, setEditingCapacity,
-    isRosterOpen, setIsRosterOpen,
-    activeRoster,
-    //handleConfirmCapacityUpdate,
-    openCapacityDialog,
-    openRosterDialog,
-    handleAddNewDate,
-    //openAddStudentDialog,
-    //handleManualAdd
-  } = useSchedules();
+export function SchedulesTab({ schedules, fetchSchedules }: ScheduleProp) {
+
+  //input states
+  const [newDateInput, setNewDateInput] = useState("");
+  const [sessionType, setSessionType] = useState("both"); // options: both, am, pm
+  const [newAmCapacity, setNewAmCapacity] = useState(50);
+  const [newPmCapacity, setNewPmCapacity] = useState(50);
+  const [isAddDateOpen, setIsAddDateOpen] = useState(false);
+
+  //input override states
+  const [manualStudentId, setManualStudentId] = useState("");
+  const [isAddStudentOpen, setIsAddStudentOpen] = useState(false);
+  const [activeAddStudentSession, setActiveAddStudentSession] = useState<{date: string, session: 'am'|'pm'} | null>(null);
+
+  //capacity override states
+  const [isEditCapacityOpen, setIsEditCapacityOpen] = useState(false);
+  const [editingCapacity, setEditingCapacity] = useState<{date: string, session: 'am'|'pm', value: number} | null>(null);
+
+  //show student state
+  const [isRosterOpen, setIsRosterOpen] = useState(false);
+  const [activeRoster, setActiveRoster] = useState<{date: string, session: 'morning' | 'afternoon', students: any[]} | null>(null);
+
+  const openCapacityDialog = (date: string, session: 'am'|'pm', currentSlots: number) => {
+    setEditingCapacity({ date, session, value: currentSlots });
+    setIsEditCapacityOpen(true);
+  };
+
+  const openRosterDialog = (date: string, session: 'morning' | 'afternoon', students: any[]) => {
+    setActiveRoster({ date, session, students });
+    setIsRosterOpen(true);
+  };
+
+  //handle add date
+  const handleAddNewDate = async () => {
+    if (!newDateInput) return;
+    const exists = schedules.some(s => s.date === newDateInput);
+    if (exists) { alert("Date already exists!"); return; }
+
+    const amLimit = (sessionType === 'both' || sessionType === 'am') ? newAmCapacity : 0;
+    const pmLimit = (sessionType === 'both' || sessionType === 'pm') ? newPmCapacity : 0;
+
+    try {
+      const res = await adminService.addSchedule(newDateInput, amLimit, pmLimit);
+      if (res.success) {
+        alert("New schedule has been added succesfully!");
+        
+        fetchSchedules();
+
+        // reset form fields
+        setNewDateInput("");
+        setNewAmCapacity(0);
+        setNewPmCapacity(0);
+        setIsAddDateOpen(false);
+      } else {
+        alert(res.reason);
+      }
+    } catch(err) {
+      console.error("Error adding schedule", err);
+      alert("Error connecting to the server");
+    }
+  };
+
+  //TODOS:
+  /*
+   const handleConfirmCapacityUpdate = () => {
+    if (editingCapacity) {
+        setSchedules(prev => prev.map(sched => {
+            if (sched.date === editingCapacity.date) {
+                return editingCapacity.session === 'am' 
+                    ? { ...sched, amSlots: editingCapacity.value } 
+                    : { ...sched, pmSlots: editingCapacity.value };
+            }
+            return sched;
+        }));
+        setIsEditCapacityOpen(false);
+        setEditingCapacity(null);
+    }
+  }; 
+  */
+
+  //manaually add student based on id number
+  /*
+  const openAddStudentDialog = (date: string, session: 'am'|'pm') => {
+      setActiveAddStudentSession({ date, session });
+      setManualStudentId("");
+      setIsAddStudentOpen(true);
+  }
+  */
+
+  /*
+  const handleManualAdd = () => {
+    if (!manualStudentId || !activeAddStudentSession) return;
+    
+    const { date, session } = activeAddStudentSession;
+
+    setSchedules(prev => prev.map(sched => {
+        if (sched.date === date) {
+            const currentStudents = session === 'am' ? sched.amStudents : sched.pmStudents;
+            const limit = session === 'am' ? sched.amSlots : sched.pmSlots;
+            
+            if (limit === 0) { alert("This session is closed."); return sched; }
+            if (currentStudents.length >= limit) { alert("Slot full!"); return sched; }
+            
+            // i-push nato ang bag-ong student sa roster array
+            const newStudent = { id: manualStudentId, name: "Manual Added Student", status: "pending" };
+            
+            return session === 'am' 
+              ? { ...sched, amStudents: [...sched.amStudents, newStudent] } 
+              : { ...sched, pmStudents: [...sched.pmStudents, newStudent] };
+        }
+        return sched;
+    }));
+    
+    setManualStudentId(""); 
+    setIsAddStudentOpen(false);
+    setActiveAddStudentSession(null);
+  };
+  */
 
   return (
     <div className="space-y-6 animate-in fade-in slide-in-from-bottom-2 duration-500 max-w-6xl mx-auto">
@@ -185,7 +283,6 @@ export function SchedulesTab() {
                                             
                                             {/* Action Buttons */}
                                             <div className="flex gap-2 pt-2">
-                                                {/* Kani tong gi request ni koi nga list sa students para transparency */}
                                                 <Button 
                                                     variant="outline" 
                                                     size="sm" 
@@ -235,11 +332,11 @@ export function SchedulesTab() {
                             <TabsTrigger value="pending" className="text-amber-700 data-[state=active]:bg-amber-50 data-[state=active]:text-amber-800">Pending/Missed</TabsTrigger>
                         </TabsList>
 
-                        {/* nag loop ko dre sa tabs para dli cgeg copy paste sa UI elements sa sulod */}
+                        {/* map result based on status */}
                         {[
                             { value: 'all', filterFn: () => true },
-                            { value: 'attended', filterFn: (s: any) => s.status === 'attended' },
-                            { value: 'pending', filterFn: (s: any) => s.status !== 'attended' },
+                            { value: 'attended', filterFn: (s: any) => s.status === 'ATTENDED' },
+                            { value: 'pending', filterFn: (s: any) => s.status !== 'ATTENDED' },
                         ].map(tab => (
                             <TabsContent key={tab.value} value={tab.value} className="mt-0">
                                 <ScrollArea className="h-[400px] pr-4">
@@ -247,14 +344,13 @@ export function SchedulesTab() {
                                         {activeRoster?.students.filter(tab.filterFn).map((student, idx) => (
                                             <div key={idx} className="flex justify-between items-center p-3 rounded-xl border border-stone-100 hover:bg-stone-50 transition-colors">
                                                 <div className="flex flex-col">
-                                                    <span className="font-bold text-sm text-stone-800">{student.name}</span>
-                                                    <span className="text-xs text-stone-400 font-mono">{student.id}</span>
+                                                    <span className="font-bold text-sm text-stone-800">{`${student.student.first_name} ${student.student.last_name}`}</span>
+                                                    <span className="text-xs text-stone-400 font-mono">{student.student_number}</span>
                                                 </div>
                                                 
-                                                {/* I-display ang status gamit ang badges */}
-                                                {student.status === 'attended' && <Badge className="bg-green-100 text-green-700 hover:bg-green-100 border-0"><CheckCircle2 className="w-3 h-3 mr-1"/> Attended</Badge>}
-                                                {student.status === 'pending' && <Badge className="bg-amber-100 text-amber-700 hover:bg-amber-100 border-0"><Clock className="w-3 h-3 mr-1"/> Pending</Badge>}
-                                                {student.status === 'missed' && <Badge className="bg-red-100 text-red-700 hover:bg-red-100 border-0"><XCircle className="w-3 h-3 mr-1"/> No Show</Badge>}
+                                                {/* display status using badges */}
+                                                {student.student.StudentAuth.status === 'ATTENDED' && <Badge className="bg-green-100 text-green-700 hover:bg-green-100 border-0"><CheckCircle2 className="w-3 h-3 mr-1"/> Attended</Badge>}
+                                                {student.student.StudentAuth.status === 'BOOKED' && <Badge className="bg-amber-100 text-amber-700 hover:bg-amber-100 border-0"><Clock className="w-3 h-3 mr-1"/> Pending</Badge>}
                                             </div>
                                         ))}
